requirements.txt
apache-airflow-providers-github==2.1.0
apache-airflow-providers-google==6.3.0